Move auth module to common
