+++ src/common/auth/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
-import {ConfigService} from '../common/config/config.service';
+import {ConfigService} from '../config/config.service';
 
 @Injectable()
 export class BearerAuthGuard extends AuthGuard('bearer') {
