+++ src/hash/hash.module.ts
@@ -5,7 +5,7 @@ import { ConfigModule } from '../common/config/config.module';
 import { NodeModule } from '../node/node.module';
 import { StorageModule } from '../storage/storage.module';
 import { EncoderModule } from '../common/encoder/encoder.module';
-import { AuthModule } from '../auth/auth.module';
+import { AuthModule } from '../common/auth/auth.module';
 import { HashService } from './hash.service';
 import { HashListenerService } from './hash-listener.service';
 import { EmitterModule } from '../emitter/emitter.module';
